+++ tests/common/test_system.py
@@ -16,17 +16,17 @@ from tests import AsyncIterator, AnyInstance
 
 class FindProcessByNameTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 # xpto ', b' 123 # abc '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 xpto ', b' 123 abc '])))
     async def test__make_exact_comparisson(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_name(re.compile('^abc$'))
         self.assertIsNotNone(proc_found)
         self.assertEqual(123, proc_found[0])
         self.assertEqual('abc', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%c" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,comm -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
     @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', side_effect=[
-        MagicMock(stdout=AsyncIterator([b'456#xpto', b'123#abc'])),
-        MagicMock(stdout=AsyncIterator([b'456#xpto', b'123#abc']))])
+        MagicMock(stdout=AsyncIterator([b'456 xpto', b'123 abc'])),
+        MagicMock(stdout=AsyncIterator([b'456 xpto', b'123 abc']))])
     async def test__make_regex_comparisson(self, create_subprocess_shell: AsyncMock):
         regex_no_match = re.compile('^.+b$')
 
@@ -40,34 +40,34 @@ class FindProcessByNameTest(IsolatedAsyncioTestCase):
 
         self.assertEqual(2, create_subprocess_shell.await_count)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'123#ac', b'456#ab'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'123 ac', b'456 ab'])))
     async def test__return_first_last_match_when_last_match_is_false(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_name(re.compile('^a.+$'))
         self.assertIsNotNone(proc_found)
         self.assertEqual(123, proc_found[0])
         self.assertEqual('ac', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%c" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,comm -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'456#ab', b'123#ac'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'456 ab', b'123 ac'])))
     async def test__return_last_match_when_last_match_is_true(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_name(re.compile('^a.+$'), last_match=True)
         self.assertIsNotNone(proc_found)
         self.assertEqual(456, proc_found[0])
         self.assertEqual('ab', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%c" -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,comm -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
 
 class FindProcessByCommandTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 # /bin/xpto ', b' 123 # /opt/abc '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 /bin/xpto ', b' 123 /opt/abc '])))
     async def test__make_exact_comparisson(self, create_subprocess_shell: MagicMock):
         proc_found = await system.find_process_by_command({re.compile('^/opt/abc$')})
         self.assertIsNotNone(proc_found)
         self.assertEqual(123, proc_found[0])
         self.assertEqual('/opt/abc', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 # /bin/xpto ', b' 123 # /opt/abc '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 456 /bin/xpto ', b' 123 /opt/abc '])))
     async def test__match_for_one_of_the_specified_patterns(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_command({re.compile('^/def$'), re.compile('^/opt/abc$')})
         self.assertIsNotNone(proc_found)
@@ -76,8 +76,8 @@ class FindProcessByCommandTest(IsolatedAsyncioTestCase):
         create_subprocess_shell.assert_awaited_once()
 
     @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', side_effect=[
-        MagicMock(stdout=AsyncIterator([b' 456 # /bin/xpto', b' 123#/abcd -xpto '])),
-        MagicMock(stdout=AsyncIterator([b' 456 # /bin/xpto', b' 123#/abcd -xpto ']))
+        MagicMock(stdout=AsyncIterator([b' 456 /bin/xpto', b' 123 /abcd -xpto '])),
+        MagicMock(stdout=AsyncIterator([b' 456 /bin/xpto', b' 123 /abcd -xpto ']))
     ])
     async def test__make_regex_comparisson(self, create_subprocess_shell: AsyncMock):
         regex_no_match = re.compile('^/.+c$')
@@ -92,47 +92,47 @@ class FindProcessByCommandTest(IsolatedAsyncioTestCase):
 
         self.assertEqual(2, create_subprocess_shell.await_count)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'456#/bin/x', b'123#/bin/a'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'456 /bin/x', b'123 /bin/a'])))
     async def test__return_last_match_when_last_match_is_true(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_command({re.compile('^/bin/.+$')}, last_match=True)
         self.assertIsNotNone(proc_found)
         self.assertEqual(456, proc_found[0])
         self.assertEqual('/bin/x', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'123#/bin/a', b'456#/bin/x'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'123 /bin/a', b'456 /bin/x'])))
     async def test__return_first_match_when_last_match_is_false(self, create_subprocess_shell: AsyncMock):
         proc_found = await system.find_process_by_command({re.compile('^/bin/.+$')}, last_match=False)
         self.assertIsNotNone(proc_found)
         self.assertEqual(123, proc_found[0])
         self.assertEqual('/bin/a', proc_found[1])
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
 
 class FindChildrenTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.async_syscall', return_value=(0, "789#101\n776#12\n"))
+    @patch(f'{__app_name__}.common.system.async_syscall', return_value=(0, "789 101\n776 12\n"))
     async def test__empty_list_when_no_children_is_found(self, async_syscall: Mock):
         children = await system.find_children({123, 456})
         self.assertEqual([], children)
 
-        async_syscall.assert_called_once_with('ps -Ao "%P#%p" -ww --no-headers')
+        async_syscall.assert_called_once_with('ps -Ao ppid,pid -ww --no-headers')
 
     @patch(f'{__app_name__}.common.system.async_syscall')
     async def test__children_sorted_by_the_deepest_on_the_tree(self, async_syscall: Mock):
         pid_tree = '''
-          1#  11 
-          1#  12 
-          11# 111 
-          11# 112 
-          2#  22 
-          2#  23 
-          23#  233 
-          233#  444 
+          1 11 
+          1 12 
+          11 111 
+          11 112 
+          2 22 
+          2 23 
+          23 233 
+          233 444 
         '''
         async_syscall.return_value = (0, pid_tree)
         children = await system.find_children({1, 2})
-        async_syscall.assert_called_once_with('ps -Ao "%P#%p" -ww --no-headers')
+        async_syscall.assert_called_once_with('ps -Ao ppid,pid -ww --no-headers')
         self.assertIsNotNone(children)
         self.assertEqual(8, len(children))
         self.assertEqual(444, children[0])
@@ -142,56 +142,56 @@ class FindChildrenTest(IsolatedAsyncioTestCase):
 
 class FindPidsByNameTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 # a ', b' 2 # b ', b' 3 # a ', b' 4 # b', b' 5 # c '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 a ', b' 2 b ', b' 3 a ', b' 4 b', b' 5 c '])))
     async def test__must_return_the_first_match_when_last_match_is_not_defined(self, create_subprocess_shell: AsyncMock):
         res = await find_pids_by_names({'a', 'b', 'c', 'd'})
 
         self.assertEqual({'a': 1, 'b': 2, 'c': 5}, res)
 
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%c" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,comm -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'5#c', b'4#b', b'3#a', b'2#b', b'1#a'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b'5  c', b'4  b', b'3  a', b'2 b', b'1 a'])))
     async def test__must_return_the_last_match_when_last_match_is_true(self, create_subprocess_shell: AsyncMock):
         res = await find_pids_by_names({'a', 'b', 'c'}, last_match=True)
 
         self.assertEqual({'a': 3, 'b': 4, 'c': 5}, res)
 
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%c" -ww --no-headers --sort=-pid',
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,comm -ww --no-headers --sort=-pid',
                                                          stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
                                                          stdin=subprocess.DEVNULL)
 
 
 class FindCommandsByPidsTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 # /bin/a ', b'  2 # /bin/b  ', b' 3 # /bin/c ', b' 4 # /bin/d -xpto '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 /bin/a ', b'  2 /bin/b  ', b' 3 /bin/c ', b' 4 /bin/d -xpto '])))
     async def test__must_return_commands_by_informed_pids(self, create_subprocess_shell: AsyncMock):
         res = await find_commands_by_pids({2, 4, 7})
         self.assertEqual({2: '/bin/b', 4: '/bin/d -xpto'}, res)
 
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers',
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers',
                                                          stdout=subprocess.PIPE, stderr=subprocess.STDOUT,
                                                          stdin=subprocess.DEVNULL)
 
 
 class FindProcessesByCommandTest(IsolatedAsyncioTestCase):
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 # /bin/a ', b' 2 # /bin/b ', b' 3 # /bin/a ', b' 4 # /bin/c', b' 5 # /bin/b ', b' 6 # /bin/d'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 /bin/a ', b' 2 /bin/b ', b' 3 /bin/a ', b' 4 /bin/c', b' 5 /bin/b ', b' 6 /bin/d'])))
     async def test__must_return_first_matches_when_parameter_not_defined(self, create_subprocess_shell: AsyncMock):
         res = await find_processes_by_command({'/bin/a', '/bin/b', '/bin/d'})
         self.assertEqual({'/bin/a': 1, '/bin/b': 2, '/bin/d': 6}, res)
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 # /bin/a ', b' 2 # /bin/b ', b' 3 # /bin/a ', b' 4 # /bin/c', b' 5 # /bin/b ', b' 6 # /bin/d'])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 1 /bin/a ', b' 2 /bin/b ', b' 3 /bin/a ', b' 4  /bin/c', b' 5  /bin/b ', b' 6  /bin/d'])))
     async def test__must_return_first_matches_when_parameter_set_to_false(self, create_subprocess_shell: AsyncMock):
         res = await find_processes_by_command({'/bin/a', '/bin/b', '/bin/d'}, last_match=False)
         self.assertEqual({'/bin/a': 1, '/bin/b': 2, '/bin/d': 6}, res)
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
-    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 6 # /bin/d', b' 5 # /bin/b ',  b' 4 # /bin/c', b' 3 # /bin/a ', b' 2 # /bin/b ',  b' 1 # /bin/a '])))
+    @patch(f'{__app_name__}.common.system.asyncio.create_subprocess_shell', return_value=MagicMock(stdout=AsyncIterator([b' 6  /bin/d', b' 5  /bin/b ',  b' 4  /bin/c', b' 3  /bin/a ', b' 2  /bin/b ',  b' 1  /bin/a '])))
     async def test__must_return_last_matches_when_parameter_set_to_true(self, create_subprocess_shell: AsyncMock):
         res = await find_processes_by_command({'/bin/a', '/bin/b', '/bin/d'}, last_match=True)
         self.assertEqual({'/bin/a': 3, '/bin/b': 5, '/bin/d': 6}, res)
-        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao "%p#%a" -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
+        create_subprocess_shell.assert_awaited_once_with(cmd='ps -Ao pid,args -ww --no-headers --sort=-pid', stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.DEVNULL)
 
 
 class MapProcessByParentTest(IsolatedAsyncioTestCase):
@@ -199,17 +199,17 @@ class MapProcessByParentTest(IsolatedAsyncioTestCase):
     @patch(f"{__app_name__}.common.system.async_syscall")
     async def test_map_processes_by_parent(self, async_syscall: AsyncMock):
         async_syscall.return_value = (0, """
-        1411#    5202# reaper
-        5202#    5203# pv-bwrap
-        5203#    5286# pressure-vessel
-        5286#    7925# python3
-        5286#    8017# wineserver
-        5286#    8708# Game.exe
-        5286#    8747# Game-Win64-Shi
+        1411    5202 reaper
+        5202    5203 pv-bwrap
+        5203    5286 pressure-vessel
+        5286    7925 python3
+        5286    8017 wineserver
+        5286    8708 Game.exe
+        5286    8747 Game-Win64-Shi
         """)
 
         proc_map = await map_processes_by_parent()
-        async_syscall.assert_awaited_with(f'ps -Ao "%P#%p#%c" -ww --no-headers')
+        async_syscall.assert_awaited_with("ps -Ao ppid,pid,comm -ww --no-headers")
         self.assertEqual(4, len(proc_map))
 
         expected = {1411: {(5202, "reaper")},